+++ include/parser.h
@@ -1,0 +1,11 @@
+#ifndef __PARSER_H__
+#define __PARSER_H__
+
+#include <Arduino.h>
+#include "motor.h"
+#include "gcode_comands.h"
+#include "temperature.h"
+
+void parsing();
+
+#endif // __PARSER_H__
