+++ include/temperature.h
@@ -1,0 +1,13 @@
+#ifndef __TEMPERATURE_H__
+#define __TEMPERATURE_H__
+
+struct {
+  double SlotTemperature;
+  double TargetTemperature;
+
+  bool isHeaterOn = 0;
+  bool isOnTarget = 0;
+} Temperature;
+
+
+#endif // __TEMPERATURE_H__
