+++ src/cord.json
@@ -1,14 +1,14 @@
-"tube-holder-coordinates" : {
-    "1":  "2:2",
-    "2":  "2:32",
-    "3":  "2:62",
-    "4":  "2:92",
-    "5":  "2:122",
-    "6":  "2:152",
+"tube-holder-coordinates": {
+    "1": "2:2",
+    "2": "2:32",
+    "3": "2:62",
+    "4": "2:92",
+    "5": "2:122",
+    "6": "2:152",
 
-    "7":  "35:2",
-    "8":  "35:32",
-    "9":  "35:62",
+    "7": "35:2",
+    "8": "35:32",
+    "9": "35:62",
     "10": "35:92",
     "11": "35:122",
     "12": "35:152",
